Corrected 64bit typo in the .bat. Added the first draft of the PowerShell script.
Detects 32/64 bitness. Doesn't check for the zip existing, overwrites it
instead. Skips 7z and curl, but only works for PowerShell 3.0 and up
(installed on 7.1 by default, available on Vista and up). Workarounds
for it working on XP are possible, but not pretty, hence not implemented
- up for discussion.

Will attempt to add the relevant section to the README later.

Future versions should add: MD5 checksums, automatic change of p2p.conf.